+++ src/lib.rs
@@ -20,6 +20,7 @@ pub mod computer_cap;
 pub mod fs_cap;
 pub mod git_tui;
 pub mod http;
+pub mod http_cap;
 pub mod interp;
 pub mod ir;
 pub mod js;